+++ spec/ckb/rpc_spec.rb
@@ -55,8 +55,14 @@ RSpec.describe CKB::RPC do
   end
 
   it "get live cell" do
-    cells = rpc.get_cells_by_lock_hash(lock_hash, '0', '100')
-    result = rpc.get_live_cell(cells[0][:out_point])
+    out_point = {
+      block_hash: nil,
+      cell: {
+        tx_hash: "0x45d086fe064ada93b6c1a6afbfd5e441d08618d326bae7b7bbae328996dfd36a",
+        index: "0"
+      }
+    }
+    result = rpc.get_live_cell(out_point)
     expect(result).not_to be nil
   end
 
@@ -112,5 +118,34 @@ RSpec.describe CKB::RPC do
       result = rpc.get_transactions_by_lock_hash(lock_hash, 0, 10)
       expect(result).not_to be nil
     end
+  end
+
+  it "get block header" do
+    block_hash = rpc.get_block_hash("1")
+    result = rpc.get_header(block_hash)
+    expect(result[:number].to_i > 0).to be true
+  end
+
+  it "get block header by number" do
+    block_number = "1"
+    result = rpc.get_header_by_number(block_number)
+    expect(result[:number]).to eq block_number
+  end
+
+  it "get block reward by block hash" do
+    block_hash = rpc.get_block_hash("1")
+    result = rpc.get_cellbase_output_capacity_details(block_hash)
+    expect(result).not_to be nil
+  end
+
+  it "set ban" do
+    params = ["192.168.0.2", "insert", "1840546800000", true, "test set_ban rpc"]
+    result = rpc.set_ban(*params)
+    expect(result).to be nil
+  end
+
+  it "get banned addresses" do
+    result = rpc.get_banned_addresses
+    expect(result).not_to be nil
   end
 end